Tahmid Chowdhury With JavaScript (If_else)
The if...else statement in JavaScript allows for conditional execution of code. It executes a block of code if a specified condition is true, and optionally executes another block if the condition is false. 
Syntax
JavaScript

if (condition) {
  // Code to execute if condition is true
} else {
  // Code to execute if condition is false
}
The else block is optional. If omitted, no code is executed when the condition is false.
Example
JavaScript

let age = 20;

if (age >= 18) {
  console.log("You are an adult.");
} else {
  console.log("You are a minor.");
}
// Output: You are an adult.
else if statement
Multiple conditions can be chained using the else if statement.
JavaScript

let score = 75;

if (score >= 90) {
  console.log("Grade A");
} else if (score >= 80) {
  console.log("Grade B");
} else if (score >= 70) {
  console.log("Grade C");
} else {
  console.log("Grade D");
}
// Output: Grade C
Ternary operator
A shorthand for the if...else statement is the ternary operator.
JavaScript

let isRaining = true;
let message = isRaining ? "Take an umbrella" : "Enjoy the sun";
console.log(message);
// Output: Take an umbrella