Tahmid Chowdhury With String Method ES6
Advanced String Methods:
match(): This method retrieves the result of matching a string against a regular expression. It returns an array containing the matched results or null if no matches are found. If the regular expression has the g (global) flag, it returns all matches; otherwise, it returns only the first match

search(): This method tests a string for a match against a regular expression and returns the index of the first match. If no match is found, it returns -1. 

replace() and replaceAll(): These methods are used to replace occurrences of a substring or a pattern (using a regular expression) within a string.
replace() replaces only the first occurrence of a substring or the first match of a regular expression (unless the global flag g is used in the regex).
replaceAll() replaces all occurrences of a substring or all matches of a regular expression.

split(): This method divides a string into an ordered list of substrings and returns them as an array. The division is based on a specified separator (which can be a string or a regular expression).

includes(), startsWith(), endsWith(): These methods are used for checking the presence or position of a substring within a string.
includes() checks if a string contains a specified substring, returning true or false.
startsWith() checks if a string begins with a specified substring.
endsWith() checks if a string ends with a specified substring.